feat: refactor config loading and add hostinfo and git packages; implement GPU info retrieval and related tests
