+++ config_alias.go
@@ -1,0 +1,5 @@
+package main
+
+import configpkg "manifold/internal/config"
+
+type Config = configpkg.Config
