+++ internal/config/config.go
@@ -1,6 +1,6 @@
 // manifold/config.go
 
-package main
+package config
 
 import (
 	"fmt"
