+++ internal/git/git_test.go
@@ -1,4 +1,4 @@
-package main
+package git
 
 import (
 	"testing"
